backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pyjwt==2.8.0
bcrypt==3.2.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart